feat: [DZ] rename COD Confirmation to Cash Confirmation and update related model
